Ensure the method does not shadow outer scope references.
We can check to see whether the function has a reference with the same identifier as the method name. If it does, we go through some extra work to ensure that the outer reference is resolved and not the method itself. This is all to preserve the function name from the method.